minchan task1

+++ src/Expansion.cpp
@@ -4,7 +4,7 @@
 
 #include "Expansion.hpp"
 
-Expansion::Expansion(int8_t const piston_port): m_piston{ piston_port } {}
+Expansion::Expansion(int8_t const piston_port): m_piston{ static_cast<uint8_t>(piston_port) } {}
 
 void Expansion::trigger() {
 	m_count += 1;
